server/pyproject.toml
[project]
name = "projex_server"
version = "1.0"
description = "Projex - Project Management Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.118.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "fakeredis>=2.32.0",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]
